Refactor get_key function and deprecate MediaPlayer:Get
`get_key` is used as `__index` function on the player object. The function has
been refactored to be a bit more readable and efficient.

Using the name of the property as a getter directly is sufficient, so `Get` is
a bit useless. The function is kept for backwards compatibility, but is not
used internally any more.